Add files via upload
This is my proposal for the PhD research programming at Victoria University. It is based on the wide range researches on the gaps in healthcare sector in Australia. This research will help to manage the risk in cardivascular disease through prevention and detection and reduce mortality and morbidity caused by the poor lifestyle and awareness in Austalia.